{en,de}crypt.go: {En,De}cryptFile: Correctly use `dest` and `overwrite` args

+++ encrypt.go
@@ -91,8 +91,15 @@ func EncryptFile(filename string, key *[32]byte, dest string, overwrite bool) (c
 	// TODO: Make the name and location of resulting encrypted
 	// file configurable with `-o <outfile>` option or similar
 
-	// Save encrypted file with ".minileap" extension appended
-	cipherFilename = filename + MiniLeapFileExtensionIncludingDot
+	cipherFilename = dest
+	if dest == "" {
+		// Save encrypted file with ".minileap" extension appended
+		cipherFilename = filename + MiniLeapFileExtensionIncludingDot
+	}
+
+	if FileExists(cipherFilename) && !overwrite {
+		return cipherFilename, fmt.Errorf("Encrypted file `%s` already exists and you've chosen not to overwrite existing files!", cipherFilename)
+	}
 
 	cipherFile, err := os.Create(cipherFilename)
 	if err != nil {
